requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deepsearch-glm==0.26.1
distro==1.9.0
docling==2.7.0
docling-core==2.4.1
docling-ibm-models==2.0.6
docling-parse==2.1.2
docutils==0.21.2
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
filetype==1.2.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jiter==0.7.1
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.14.1
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
openai==1.55.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
pyarrow==16.1.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pypdfium2==4.30.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==307
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
Rtree==1.3.0
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.40.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
XlsxWriter==3.2.0
